separate packages in data domain and presentation
