+++ domain/entity/Food.java
@@ -1,4 +1,4 @@
-package entity;
+package domain.entity;
 
 import java.util.List;
