+++ domain/entity/Question.java
@@ -1,4 +1,4 @@
-package entity;
+package domain.entity;
 
 import java.util.ArrayList;
 import java.util.List;
