+++ presentation/Home.java
@@ -1,9 +1,8 @@
-import entity.Answer;
-import entity.Food;
-import entity.FoodFormData;
-import entity.Question;
-import utils.DataProvider;
-import utils.SwingUtils;
+package presentation;
+
+import domain.entity.*;
+import data.DataProvider;
+import presentation.SwingUtils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,44 +16,43 @@ public class Home {
 
     static List<Food> filtratedFoodList = new ArrayList();
     static List<Answer> roundAnswers = new ArrayList();
+
     public static void main(String[] args) {
         setInitialData();
         gameLoop();
     }
-    public static void calculatePriority(){
-        questions.parallelStream().forEach( question -> question.calculateWeight());
+
+    public static void calculatePriority() {
+        questions.parallelStream().forEach(question -> question.calculateWeight());
     }
 
-    public static void gameLoop(){
+    public static void gameLoop() {
         filtratedFoodList = new ArrayList(foodList);
         roundAnswers = new ArrayList();
         SwingUtils.showGreetings();
         guessFood();
     }
 
-    public static void  guessFood(){
+    public static void guessFood() {
         var question = questionsQueue.poll();
-
         var answer = SwingUtils.askQuestion(question.getName());
-
         var givenAnswer = question.createAnswer(answer);
-            roundAnswers.add(givenAnswer);
-
+        roundAnswers.add(givenAnswer);
         filtratedFoodList = getFilterFoods(filtratedFoodList, givenAnswer);
         var foodCounter = filtratedFoodList.size();
         howManyFoods(foodCounter);
     }
 
-    public static void howManyFoods(int foodCounter){
+    public static void howManyFoods(int foodCounter) {
         if (foodCounter == 1) {
             var correctAnswer = SwingUtils.testGuess(filtratedFoodList.get(0).getName());
             if (correctAnswer) {
                 restartGame(SwingUtils.showCorrectPlayAgain());
-            }else {
+            } else {
                 askForLearn(roundAnswers);
                 restartGame(SwingUtils.showWrongPlayAgain());
             }
-        }else if(foodCounter == 0){
+        } else if (foodCounter == 0) {
             askForLearn(roundAnswers);
             restartGame(SwingUtils.showWrongPlayAgain());
         } else {
@@ -62,15 +60,15 @@ public class Home {
         }
     }
 
-    public static void setInitialData(){
+    public static void setInitialData() {
         questions = DataProvider.initialQuestions();
         foodList = DataProvider.initialFoodList(questions);
         calculatePriority();
         questionsQueue.addAll(questions);
     }
 
-    public static void restartGame(Boolean playAgain){
-        if(playAgain) {
+    public static void restartGame(Boolean playAgain) {
+        if (playAgain) {
             questionsQueue = new PriorityQueue(new QuestionComparator());
             calculatePriority();
             questionsQueue.addAll(questions);
@@ -79,13 +77,13 @@ public class Home {
         }
     }
 
-    public static List<Food> getFilterFoods(List<Food> foodList, Answer answer){
+    public static List<Food> getFilterFoods(List<Food> foodList, Answer answer) {
         return foodList.parallelStream()
-                .filter( food -> food.matchAnswer(answer) )
+                .filter(food -> food.matchAnswer(answer))
                 .collect(Collectors.toList());
     }
 
-    public static void askForLearn(List<Answer> roundAnswers){
+    public static void askForLearn(List<Answer> roundAnswers) {
         var newFoodName = SwingUtils.learnFood();
         var newQuestionName = SwingUtils.learnQuestion(newFoodName);
         var buttonList = SwingUtils.showOptionsToLearn(newQuestionName, foodList);
@@ -95,18 +93,18 @@ public class Home {
         foodListLearnNewQuestion(formData, newQuestion);
     }
 
-    public static void learnNewFood(String newFoodName,Question newQuestion ,List<Answer> answers){
+    public static void learnNewFood(String newFoodName, Question newQuestion, List<Answer> answers) {
         answers.add(newQuestion.createAnswer(true));
         foodList.add(new Food(newFoodName, answers));
     }
 
-    public static Question learnNewQuestion(String newQuestionName){
+    public static Question learnNewQuestion(String newQuestionName) {
         var newQuestion = new Question(newQuestionName);
         questions.add(newQuestion);
         return newQuestion;
     }
 
-    public static void foodListLearnNewQuestion(List<FoodFormData> formData, Question newQuestion){
+    public static void foodListLearnNewQuestion(List<FoodFormData> formData, Question newQuestion) {
         formData.forEach(form -> {
                     foodList.forEach(food -> {
                                 if (food.getName() == form.getName())
